feat(#379): refactored
