+++ apps/web/src/app/[lang]/network/_components/network-selected.tsx
@@ -1,55 +1,44 @@
 'use client';
 
 import { Category, Space } from '@hypha-platform/core/client';
-import { SpaceGroupSlider, SpaceSearch } from '@hypha-platform/epics';
+import { CategorySpaces } from '@hypha-platform/epics';
 import { Locale } from '@hypha-platform/i18n';
-import { Button } from '@hypha-platform/ui';
-import { PlusIcon } from '@radix-ui/react-icons';
+import { ChevronLeftIcon } from '@radix-ui/react-icons';
 import { Text } from '@radix-ui/themes';
 import Link from 'next/link';
-import { getDhoPathGovernance } from '../../dho/[id]/@tab/governance/constants';
 
 export function NetworkSelected({
   lang,
   spaces,
-  uniqueCategories,
+  categories,
 }: {
   lang: Locale;
   spaces: Space[];
-  uniqueCategories: Category[];
+  categories: Category[];
 }) {
-  const getHref = (id: string) => {
-    return getDhoPathGovernance(lang, id);
-  };
-
   return (
     <>
-      <Text className="text-9 text-center flex mb-8">
-        Explore the Hypha Network
-      </Text>
-      <div className="flex justify-center">
-        <SpaceSearch />
-        <Link href={`/${lang}/network/create`} scroll={false}>
-          <Button className="ml-2">
-            <PlusIcon className="mr-2" />
-            Create Space
-          </Button>
+      <div className="mb-6 flex items-center">
+        <Link
+          href={`/${lang}/network`}
+          className="cursor-pointer flex items-center"
+        >
+          <ChevronLeftIcon width={16} height={16} />
+          <Text className="text-sm">My Spaces</Text>
         </Link>
+        <Text className="text-sm text-gray-400 ml-1 font-medium capitalize">
+          {' '}
+          / {categories[0]}
+        </Text>
       </div>
-      {uniqueCategories.map((category) => (
-        <SpaceGroupSlider
-          key={category}
-          spaces={spaces.filter((space) =>
-            space.categories?.includes(category),
-          )}
-          type={category}
-          getHref={getHref}
-        />
-      ))}
-      <SpaceGroupSlider
-        spaces={spaces.filter((space) => space.categories?.length === 0)}
-        type={'No Category'}
-        getHref={getHref}
+      <Text className="text-9 text-center flex mb-8">
+        Category Spaces
+      </Text>
+      {/* TODO: search input */}
+      <CategorySpaces
+        lang={lang}
+        spaces={spaces}
+        categories={categories}
       />
     </>
   );
